+++ dds_web/api/__init__.py
@@ -79,12 +79,15 @@ api.add_resource(user.Users, "/users", endpoint="users")
 
 # Super Admins ###################################################################### Super Admins #
 
+api.add_resource(superadmin_only.SetMaintenance, "/maintenance", endpoint="maintenance")
 api.add_resource(superadmin_only.AllUnits, "/unit/info/all", endpoint="all_units")
 api.add_resource(superadmin_only.MOTD, "/motd", endpoint="motd")
+api.add_resource(superadmin_only.SendMOTD, "/motd/send", endpoint="send_motd")
 api.add_resource(superadmin_only.FindUser, "/user/find", endpoint="find_user")
 api.add_resource(
     superadmin_only.ResetTwoFactor, "/user/totp/deactivate", endpoint="reset_user_hotp"
 )
+api.add_resource(superadmin_only.AnyProjectsBusy, "/proj/busy/any", endpoint="projects_busy_any")
 
 # Invoicing ############################################################################ Invoicing #
 api.add_resource(user.ShowUsage, "/usage", endpoint="usage")